% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muirpkgs-package.R
\docType{package}
\name{muirpkgs-package}
\alias{muirpkgs}
\alias{muirpkgs-package}
\title{muirpkgs: Attach Muir Lab R Packages}
\description{
Attaches CRAN and GitHub packages commonly used in
    the Muir lab.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/muir-lab/muirpkgs}
  \item Report bugs at \url{https://github.com/muir-lab/muirpkgs}
}

}
\author{
\strong{Maintainer}: Christopher Muir \email{cdmuir@hawaii.edu} (\href{https://orcid.org/0000-0003-2555-3878}{ORCID})

}
\keyword{internal}
